+++ Backend/Models/Proizvod.cs
@@ -5,7 +5,7 @@ namespace Backend.Models
     public class Proizvod: Entitet
     {
         public string? Naziv { get; set; }
-        public decimal? Tezina { get; set; }
+        public int? Tezina { get; set; }
         public decimal? Cijena { get; set; }
         [ForeignKey("kategorija")]
         public required Kategorija Kategorija { get; set; }
